DAOCustomer is implemented. It has three update and two find operations. IFDAOCustomer is updated, I have just changed the headers a bit, to give it more clarity. I hope it is so.
I will work more on DAOCustomer because I have to do more on the find operations and check the update operations if they're implemented appropriately. 

Because, now, a customer instance is made, but it is only populated with data from the Customer table. Customer it must be populated with data from Customer, Person and Location tables.