DAOCustomer and DAOPerson (and their interfaces) are updated. Now when you insert/update Customer, first DAOPerson is made which adds data in Location and Person tables, and then into the Customer table;
When you delete the order is reversed; First you delete the Customer from the db then the Person; Nothing is deleted in Location, because I think there's no point of it. You can have several customers with same zipCode and from same country and city;

However, in DAOPerson, the method insertLocation will be updated so that it first perform check in the db whether there is already such a location inserted. If there isn't such a location it will be inserted in the table;

//stefo;